BlobContrast/BlobMorphology/BlobVariance/LoG: fixed all spot coordinates to be float[], so that they compile without errors, but they still don't use physical vs pixel coordinates correctly.
I need to be able to test the other analyzers as I modify them, so I had to make these at least compile by fixing types.

+++ src-plugins/Embryo_Tracker/fiji/plugin/nperry/features/BlobContrast.java
@@ -32,7 +32,7 @@ public class BlobContrast <T extends RealType<T>> extends IndependentFeatureAnal
 
 	@Override
 	public void process(Spot spot) {
-		double contrast = getContrast(spot, diam);
+		float contrast = getContrast(spot, diam);
 		spot.addFeature(FEATURE, Math.abs(contrast));
 	}
